# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: GreenSync Lawn Care SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: assignedUserId
          in: query
          schema:
            type: integer
        - name: customerId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Appointment list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentListResponse"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentRequest"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentDetail"
    put:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /appointments/{id}/complete:
    post:
      operationId: completeAppointment
      tags: [appointments]
      summary: Mark appointment as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                completionNotes:
                  type: string
      responses:
        "200":
          description: Appointment completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"

  # ─── RECURRING PLANS ─────────────────────────────────────────────────────────
...
  /routes:
    get:
      operationId: listRoutes
      tags: [routes]
      summary: List routes
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Route list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteListResponse"
    post:
      operationId: createRoute
      tags: [routes]
      summary: Create a route
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRouteRequest"
      responses:
        "201":
          description: Route created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
  /routes/{id}:
    get:
      operationId: getRoute
      tags: [routes]
      summary: Get route with stops
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Route detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteDetail"
    put:
      operationId: updateRoute
      tags: [routes]
      summary: Update route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRouteRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
    delete:
      operationId: deleteRoute
      tags: [routes]
      summary: Delete route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # ─── AUTH SCHEMAS ──────────────────────────────────────────────────────────
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        addressLine1:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        zip:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        leadSource:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, companyId, firstName, lastName, createdAt]
...
    Property:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        customerId:
          type: integer
        propertyName:
          type: string
          nullable: true
        addressLine1:
          type: string
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        zip:
          type: string
          nullable: true
        gateNotes:
          type: string
          nullable: true
        yardSize:
          type: string
          nullable: true
        propertyNotes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, companyId, customerId, addressLine1, createdAt]
...
    Service:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        durationMinutes:
          type: integer
          nullable: true
        basePrice:
          type: number
          nullable: true
        isActive:
          type: boolean
        category:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, companyId, name, isActive, createdAt]
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        customerId:
          type: integer
        propertyId:
          type: integer
          nullable: true
        serviceId:
          type: integer
          nullable: true
        assignedUserId:
          type: integer
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, in_progress, completed, cancelled, no_show]
        scheduledStart:
          type: string
          format: date-time
        scheduledEnd:
          type: string
          format: date-time
          nullable: true
        price:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        internalNotes:
          type: string
          nullable: true
        completionNotes:
          type: string
          nullable: true
        reminderSent:
          type: boolean
        customerName:
          type: string
          nullable: true
        serviceName:
          type: string
          nullable: true
        assignedUserName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, companyId, customerId, status, scheduledStart, reminderSent, createdAt]
    AppointmentDetail:
      type: object
      properties:
        appointment:
          $ref: "#/components/schemas/Appointment"
        customer:
          $ref: "#/components/schemas/Customer"
        property:
          $ref: "#/components/schemas/Property"
        service:
          $ref: "#/components/schemas/Service"
      required: [appointment, customer]
    AppointmentListResponse:
      type: object
      properties:
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [appointments, total, page, limit]
    CreateAppointmentRequest:
      type: object
      required: [customerId, scheduledStart, status]
      properties:
        customerId:
          type: integer
        propertyId:
          type: integer
        serviceId:
          type: integer
        assignedUserId:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, in_progress, completed, cancelled, no_show]
        scheduledStart:
          type: string
          format: date-time
        scheduledEnd:
          type: string
          format: date-time
        price:
          type: number
        notes:
          type: string
        internalNotes:
          type: string
    UpdateAppointmentRequest:
      type: object
      properties:
        customerId:
          type: integer
        propertyId:
          type: integer
        serviceId:
          type: integer
        assignedUserId:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, in_progress, completed, cancelled, no_show]
        scheduledStart:
          type: string
          format: date-time
        scheduledEnd:
          type: string
          format: date-time
        price:
          type: number
        notes:
          type: string
        internalNotes:
          type: string
        completionNotes:
          type: string

    # ─── RECURRING PLAN SCHEMAS ────────────────────────────────────────────────
...
    Route:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        name:
          type: string
          nullable: true
        routeDate:
          type: string
          format: date-time
        assignedUserId:
          type: integer
          nullable: true
        status:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        stopsCount:
          type: integer
          nullable: true
        assignedUserName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, companyId, routeDate, createdAt]
    RouteStop:
      type: object
      properties:
        id:
          type: integer
        routeId:
          type: integer
        appointmentId:
          type: integer
        stopOrder:
          type: integer
        estimatedArrival:
          type: string
          format: date-time
          nullable: true
        appointment:
          $ref: "#/components/schemas/Appointment"
      required: [id, routeId, appointmentId, stopOrder]
    RouteDetail:
      type: object
      properties:
        route:
          $ref: "#/components/schemas/Route"
        stops:
          type: array
          items:
            $ref: "#/components/schemas/RouteStop"
      required: [route, stops]
    RouteListResponse:
      type: object
      properties:
        routes:
          type: array
          items:
            $ref: "#/components/schemas/Route"
      required: [routes]
    CreateRouteRequest:
      type: object
      required: [routeDate]
      properties:
        name:
          type: string
        routeDate:
          type: string
          format: date-time
        assignedUserId:
          type: integer
        status:
          type: string
        notes:
          type: string
...